+++ lib/capybara/poltergeist/driver.rb
@@ -255,6 +255,10 @@ module Capybara::Poltergeist
       browser.remove_cookie(name)
     end
 
+    def clear_cookies
+      browser.clear_cookies
+    end
+
     def cookies_enabled=(flag)
       browser.cookies_enabled = flag
     end